Add files via upload
Introduction
The primary aim of this analysis is to explore the characteristics of the Global Intelligence chart dataset and examine how the attributes to be part of the top intelligent people globally.
Power BI was used as the primary tool for data processing and visualization in the analysis. Data Query functions were instrumental in organizing and summarizing the data, allowing for a structured exploration of key factors such as the top intelligent person, top IQ country, IQ work done and average IQ which were used to derived insights. Additionally, a range of charts and visualizations were created to represent the findings clearly, making the analysis more accessible and interpretable for a broader audience.
Story of the Data
The data tells a compelling story about what drives The Global Intelligent Chart i.e. what is the key focus which is IQ. The data was downloaded from kaggle.com and the data was well structured, the dataset consists of 1001rows and 17columns.
Data Splitting and Preprocessing
Data Cleaning
To ensure a well-organized and user-friendly dataset, several formatting enhancements were applied using the data query in power BI.
Pre-Analysis
The objective of this research focuses on identifying top intellectual individuals worldwide based on their IQ and achievements. It aims to highlight people with exceptional knowledge and contributions in various fields.
While the expectation aims to identify the country with the highest concentration of intelligent people. Based on the findings, more educational materials will be made available in that region to support further research and knowledge development.
In-Analysis
These board reinforce the project's goal of identifying the best intellectuals and countries for knowledge advancement.
The United Kingdom (UK) ranks as the top country with the highest number of intellectuals, based on IQ data.
Physics education provides an advantage in intelligence rankings those with physics degrees dominate the list.
Quantum mechanics stands out as a major discovery contributed by highly intelligent individuals.
Nobel Prizes are the most prestigious awards granted to top intellectuals, with 1651 awards given so far.
The Dirac equation plays a crucial role in understanding neutrinos and their behavior.
Observation
This further supports the research objective of identifying global intellectual leaders and understanding the factors contributing to high intelligence.
Stephen remains the top intellectual individual globally, consistently ranking as the most intelligent person.
The UK continues to lead as the country with the highest concentration of intellectuals, boasting IQ levels reaching 2255.
Physics education significantly impacts intelligence rankings, with Ph.D. holders standing out as highly intelligent.
Quantum mechanics remains one of the greatest discoveries made by intellectuals.
The Dirac equation is a key contribution, helping describe the behavior of neutrinos such as electrons and their intrinsic spin.
Nobel Prizes continue to be the highest awards granted, with 1651 prizes awarded so far.
Recommendation
This research reinforces the importance of physics, advanced education, and key scientific discoveries in defining global intelligence. The ultimate goal is to recognize outstanding intellectuals and foster an environment that encourages innovation, learning, and excellence.
Stephen continues to rank as the most intellectual person globally, consistently leading in various intelligence metrics.
The United Kingdom (UK) remains the top country with the highest number of intellectual individuals, boasting IQ levels reaching 2255.
Physics education plays a crucial role in intelligence rankings, with Ph.D. holders leading the charts.
Quantum mechanics is recognized as one of the most significant discoveries made by top intellectuals.
The Dirac equation remains an essential contribution, helping to describe the behavior of fermions like electrons and their intrinsic spin.
Nobel Prizes continue to be the most prestigious awards granted to intellectuals, with 1651 prizes awarded so far.
Conclusion
This analysis has provided valuable insights into the performance of The Global Intelligent Chart on how individual IQ and work done are been generated.